docs: Adding extra clarification on the simbrief import in BG
Users who start using the Beginners guide might already plan or load
into the plane from the World Map using a arrival airport before
understanding or realizing that they should not do so if they intent
to import from Simbrief.

This adds a warning early in the process, in the hopes to avoid this.